requirements.txt
PyYAML
requests
json5
Flask